Split tests.rs into modules to comply with 1000-line limit
The check-file-size.rs CI script enforces a maximum of 1000 lines
per Rust source file. Split the 1674-line tests.rs into:
- tests/mod.rs (224 lines) - shared test implementations
- tests/list_tests.rs (546 lines) - linked list tests
- tests/tree_tests.rs (901 lines) - tree tests
