+++ src/tests/tree_tests.rs
@@ -1,781 +1,9 @@
-//! Comprehensive tests for 100% code coverage of platform-trees
-
-use crate::{
-    AbsoluteCircularLinkedList, AbsoluteLinkedList, IterativeSizeBalancedTree, LinkType,
-    LinkedList, RecursiveSizeBalancedTree, RelativeCircularLinkedList, RelativeLinkedList,
-};
-
-// =============================================================================
-// Test implementations
-// =============================================================================
-
-/// A simple node structure for testing linked lists
-#[derive(Debug, Clone, Copy, Default)]
-struct Node {
-    prev: usize,
-    next: usize,
-}
-
-/// A simple absolute linked list implementation for testing
-struct TestAbsoluteList {
-    nodes: Vec<Node>,
-    first: usize,
-    last: usize,
-    size: usize,
-}
-
-impl TestAbsoluteList {
-    fn new(capacity: usize) -> Self {
-        let mut nodes = Vec::with_capacity(capacity + 1);
-        // Index 0 is reserved as "null"
-        nodes.resize(capacity + 1, Node::default());
-        Self {
-            nodes,
-            first: 0,
-            last: 0,
-            size: 0,
-        }
-    }
-}
-
-impl LinkedList<usize> for TestAbsoluteList {
-    fn get_previous(&self, element: usize) -> usize {
-        self.nodes[element].prev
-    }
-
-    fn get_next(&self, element: usize) -> usize {
-        self.nodes[element].next
-    }
-
-    fn set_previous(&mut self, element: usize, previous: usize) {
-        self.nodes[element].prev = previous;
-    }
-
-    fn set_next(&mut self, element: usize, next: usize) {
-        self.nodes[element].next = next;
-    }
-}
-
-impl AbsoluteLinkedList<usize> for TestAbsoluteList {
-    fn get_first(&self) -> usize {
-        self.first
-    }
-
-    fn get_last(&self) -> usize {
-        self.last
-    }
-
-    fn get_size(&self) -> usize {
-        self.size
-    }
-
-    fn set_first(&mut self, element: usize) {
-        self.first = element;
-    }
-
-    fn set_last(&mut self, element: usize) {
-        self.last = element;
-    }
-
-    fn set_size(&mut self, size: usize) {
-        self.size = size;
-    }
-}
-
-impl AbsoluteCircularLinkedList<usize> for TestAbsoluteList {}
-
-/// A relative linked list implementation for testing (list head stored in element)
-struct TestRelativeList {
-    nodes: Vec<Node>,
-    // Store first/last/size for each "head" element
-    heads: Vec<(usize, usize, usize)>, // (first, last, size)
-}
-
-impl TestRelativeList {
-    fn new(capacity: usize) -> Self {
-        let mut nodes = Vec::with_capacity(capacity + 1);
-        nodes.resize(capacity + 1, Node::default());
-        let mut heads = Vec::with_capacity(capacity + 1);
-        heads.resize(capacity + 1, (0, 0, 0));
-        Self { nodes, heads }
-    }
-}
-
-impl LinkedList<usize> for TestRelativeList {
-    fn get_previous(&self, element: usize) -> usize {
-        self.nodes[element].prev
-    }
-
-    fn get_next(&self, element: usize) -> usize {
-        self.nodes[element].next
-    }
-
-    fn set_previous(&mut self, element: usize, previous: usize) {
-        self.nodes[element].prev = previous;
-    }
-
-    fn set_next(&mut self, element: usize, next: usize) {
-        self.nodes[element].next = next;
-    }
-}
-
-impl RelativeLinkedList<usize> for TestRelativeList {
-    fn get_first(&self, head: usize) -> usize {
-        self.heads[head].0
-    }
-
-    fn get_last(&self, head: usize) -> usize {
-        self.heads[head].1
-    }
-
-    fn get_size(&self, head: usize) -> usize {
-        self.heads[head].2
-    }
-
-    fn set_first(&mut self, head: usize, element: usize) {
-        self.heads[head].0 = element;
-    }
-
-    fn set_last(&mut self, head: usize, element: usize) {
-        self.heads[head].1 = element;
-    }
-
-    fn set_size(&mut self, head: usize, size: usize) {
-        self.heads[head].2 = size;
-    }
-}
-
-impl RelativeCircularLinkedList<usize> for TestRelativeList {}
-
-/// A tree node structure for testing `SizeBalancedTree`
-#[derive(Debug, Clone, Copy, Default)]
-struct TreeNode {
-    left: usize,
-    right: usize,
-    size: usize,
-}
-
-/// A simple `SizeBalancedTree` implementation for testing
-struct TestTree {
-    nodes: Vec<TreeNode>,
-}
-
-impl TestTree {
-    fn new(capacity: usize) -> Self {
-        let mut nodes = Vec::with_capacity(capacity + 1);
-        nodes.resize(capacity + 1, TreeNode::default());
-        Self { nodes }
-    }
-}
-
-impl RecursiveSizeBalancedTree<usize> for TestTree {
-    unsafe fn get_mut_left_reference(&mut self, node: usize) -> *mut usize {
-        &mut self.nodes[node].left
-    }
-
-    unsafe fn get_mut_right_reference(&mut self, node: usize) -> *mut usize {
-        &mut self.nodes[node].right
-    }
-
-    unsafe fn get_left_reference(&self, node: usize) -> *const usize {
-        &self.nodes[node].left
-    }
-
-    unsafe fn get_right_reference(&self, node: usize) -> *const usize {
-        &self.nodes[node].right
-    }
-
-    unsafe fn get_left(&self, node: usize) -> usize {
-        self.nodes[node].left
-    }
-
-    unsafe fn get_right(&self, node: usize) -> usize {
-        self.nodes[node].right
-    }
-
-    unsafe fn get_size(&self, node: usize) -> usize {
-        self.nodes[node].size
-    }
-
-    unsafe fn set_left(&mut self, node: usize, left: usize) {
-        self.nodes[node].left = left;
-    }
-
-    unsafe fn set_right(&mut self, node: usize, right: usize) {
-        self.nodes[node].right = right;
-    }
-
-    unsafe fn set_size(&mut self, node: usize, size: usize) {
-        self.nodes[node].size = size;
-    }
-
-    unsafe fn first_is_to_the_left_of_second(&self, first: usize, second: usize) -> bool {
-        first < second
-    }
-
-    unsafe fn first_is_to_the_right_of_second(&self, first: usize, second: usize) -> bool {
-        first > second
-    }
-}
-
-impl IterativeSizeBalancedTree<usize> for TestTree {}
-
-// =============================================================================
-// LinkType trait tests
-// =============================================================================
-
-#[cfg(test)]
-mod link_type_tests {
-    use super::*;
-
-    #[test]
-    fn test_funty_zero() {
-        assert_eq!(usize::funty(0), 0usize);
-        assert_eq!(u8::funty(0), 0u8);
-        assert_eq!(u16::funty(0), 0u16);
-        assert_eq!(u32::funty(0), 0u32);
-        assert_eq!(u64::funty(0), 0u64);
-    }
-
-    #[test]
-    fn test_funty_one() {
-        assert_eq!(usize::funty(1), 1usize);
-        assert_eq!(u8::funty(1), 1u8);
-        assert_eq!(u16::funty(1), 1u16);
-        assert_eq!(u32::funty(1), 1u32);
-        assert_eq!(u64::funty(1), 1u64);
-    }
-
-    #[test]
-    fn test_funty_various_values() {
-        for i in 0..=255u8 {
-            assert_eq!(u8::funty(i), i);
-            assert_eq!(u16::funty(i), u16::from(i));
-            assert_eq!(u32::funty(i), u32::from(i));
-            assert_eq!(u64::funty(i), u64::from(i));
-            assert_eq!(usize::funty(i), i as usize);
-        }
-    }
-}
-
-// =============================================================================
-// LinkedList trait tests
-// =============================================================================
-
-#[cfg(test)]
-mod linked_list_tests {
-    use super::*;
-
-    #[test]
-    fn test_get_set_previous() {
-        let mut list = TestAbsoluteList::new(10);
-        list.set_previous(1, 5);
-        assert_eq!(list.get_previous(1), 5);
-    }
-
-    #[test]
-    fn test_get_set_next() {
-        let mut list = TestAbsoluteList::new(10);
-        list.set_next(1, 3);
-        assert_eq!(list.get_next(1), 3);
-    }
-}
-
-// =============================================================================
-// AbsoluteLinkedList trait tests
-// =============================================================================
-
-#[cfg(test)]
-mod absolute_linked_list_tests {
-    use super::*;
-
-    #[test]
-    fn test_get_set_first() {
-        let mut list = TestAbsoluteList::new(10);
-        list.set_first(5);
-        assert_eq!(list.get_first(), 5);
-    }
-
-    #[test]
-    fn test_get_set_last() {
-        let mut list = TestAbsoluteList::new(10);
-        list.set_last(7);
-        assert_eq!(list.get_last(), 7);
-    }
-
-    #[test]
-    fn test_get_set_size() {
-        let mut list = TestAbsoluteList::new(10);
-        list.set_size(3);
-        assert_eq!(list.get_size(), 3);
-    }
-
-    #[test]
-    fn test_inc_size() {
-        let mut list = TestAbsoluteList::new(10);
-        list.set_size(5);
-        list.inc_size();
-        assert_eq!(list.get_size(), 6);
-    }
-
-    #[test]
-    fn test_dec_size() {
-        let mut list = TestAbsoluteList::new(10);
-        list.set_size(5);
-        list.dec_size();
-        assert_eq!(list.get_size(), 4);
-    }
-}
-
-// =============================================================================
-// AbsoluteCircularLinkedList trait tests
-// =============================================================================
-
-#[cfg(test)]
-mod absolute_circular_linked_list_tests {
-    use super::*;
-
-    #[test]
-    fn test_attach_as_first_empty_list() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_first(1);
-
-        assert_eq!(list.get_first(), 1);
-        assert_eq!(list.get_last(), 1);
-        assert_eq!(list.get_size(), 1);
-        assert_eq!(list.get_previous(1), 1); // Points to itself
-        assert_eq!(list.get_next(1), 1); // Points to itself
-    }
-
-    #[test]
-    fn test_attach_as_first_non_empty_list() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_first(1);
-        list.attach_as_first(2);
-
-        assert_eq!(list.get_first(), 2);
-        assert_eq!(list.get_last(), 1);
-        assert_eq!(list.get_size(), 2);
-    }
-
-    #[test]
-    fn test_attach_as_last_empty_list() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_last(1);
-
-        assert_eq!(list.get_first(), 1);
-        assert_eq!(list.get_last(), 1);
-        assert_eq!(list.get_size(), 1);
-    }
-
-    #[test]
-    fn test_attach_as_last_non_empty_list() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_last(1);
-        list.attach_as_last(2);
-
-        assert_eq!(list.get_first(), 1);
-        assert_eq!(list.get_last(), 2);
-        assert_eq!(list.get_size(), 2);
-    }
-
-    #[test]
-    fn test_attach_before() {
-        let mut list = TestAbsoluteList::new(10);
-
-        // Create list: 1 <-> 2
-        list.attach_as_first(1);
-        list.attach_as_last(2);
-
-        // Insert 3 before 2: 1 <-> 3 <-> 2
-        list.attach_before(2, 3);
-
-        assert_eq!(list.get_first(), 1);
-        assert_eq!(list.get_last(), 2);
-        assert_eq!(list.get_size(), 3);
-        assert_eq!(list.get_next(1), 3);
-        assert_eq!(list.get_next(3), 2);
-        assert_eq!(list.get_previous(2), 3);
-    }
-
-    #[test]
-    fn test_attach_before_first() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_first(1);
-        list.attach_before(1, 2);
-
-        assert_eq!(list.get_first(), 2);
-        assert_eq!(list.get_last(), 1);
-        assert_eq!(list.get_size(), 2);
-    }
-
-    #[test]
-    fn test_attach_after() {
-        let mut list = TestAbsoluteList::new(10);
-
-        // Create list: 1 <-> 2
-        list.attach_as_first(1);
-        list.attach_as_last(2);
-
-        // Insert 3 after 1: 1 <-> 3 <-> 2
-        list.attach_after(1, 3);
-
-        assert_eq!(list.get_first(), 1);
-        assert_eq!(list.get_last(), 2);
-        assert_eq!(list.get_size(), 3);
-        assert_eq!(list.get_next(1), 3);
-        assert_eq!(list.get_next(3), 2);
-    }
-
-    #[test]
-    fn test_attach_after_last() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_first(1);
-        list.attach_after(1, 2);
-
-        assert_eq!(list.get_first(), 1);
-        assert_eq!(list.get_last(), 2);
-        assert_eq!(list.get_size(), 2);
-    }
-
-    #[test]
-    fn test_detach_single_element() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_first(1);
-        list.detach(1);
-
-        assert_eq!(list.get_first(), 0);
-        assert_eq!(list.get_last(), 0);
-        assert_eq!(list.get_size(), 0);
-        assert_eq!(list.get_previous(1), 0);
-        assert_eq!(list.get_next(1), 0);
-    }
-
-    #[test]
-    fn test_detach_first_element() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_first(1);
-        list.attach_as_last(2);
-        list.attach_as_last(3);
-
-        list.detach(1);
-
-        assert_eq!(list.get_first(), 2);
-        assert_eq!(list.get_last(), 3);
-        assert_eq!(list.get_size(), 2);
-    }
-
-    #[test]
-    fn test_detach_last_element() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_first(1);
-        list.attach_as_last(2);
-        list.attach_as_last(3);
-
-        list.detach(3);
-
-        assert_eq!(list.get_first(), 1);
-        assert_eq!(list.get_last(), 2);
-        assert_eq!(list.get_size(), 2);
-    }
-
-    #[test]
-    fn test_detach_middle_element() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_first(1);
-        list.attach_as_last(2);
-        list.attach_as_last(3);
-
-        list.detach(2);
-
-        assert_eq!(list.get_first(), 1);
-        assert_eq!(list.get_last(), 3);
-        assert_eq!(list.get_size(), 2);
-        assert_eq!(list.get_next(1), 3);
-        assert_eq!(list.get_previous(3), 1);
-    }
-
-    #[test]
-    fn test_circular_structure() {
-        let mut list = TestAbsoluteList::new(10);
-
-        list.attach_as_first(1);
-        list.attach_as_last(2);
-        list.attach_as_last(3);
-
-        // Check circular structure
-        assert_eq!(list.get_next(3), 1); // Last points to first
-        assert_eq!(list.get_previous(1), 3); // First points back to last
-    }
-}
-
-// =============================================================================
-// RelativeLinkedList trait tests
-// =============================================================================
-
-#[cfg(test)]
-mod relative_linked_list_tests {
-    use super::*;
-
-    #[test]
-    fn test_get_set_first() {
-        let mut list = TestRelativeList::new(10);
-        list.set_first(1, 5);
-        assert_eq!(list.get_first(1), 5);
-    }
-
-    #[test]
-    fn test_get_set_last() {
-        let mut list = TestRelativeList::new(10);
-        list.set_last(1, 7);
-        assert_eq!(list.get_last(1), 7);
-    }
-
-    #[test]
-    fn test_get_set_size() {
-        let mut list = TestRelativeList::new(10);
-        list.set_size(1, 3);
-        assert_eq!(list.get_size(1), 3);
-    }
-
-    #[test]
-    fn test_inc_size() {
-        let mut list = TestRelativeList::new(10);
-        list.set_size(1, 5);
-        list.inc_size(1);
-        assert_eq!(list.get_size(1), 6);
-    }
-
-    #[test]
-    fn test_dec_size() {
-        let mut list = TestRelativeList::new(10);
-        list.set_size(1, 5);
-        list.dec_size(1);
-        assert_eq!(list.get_size(1), 4);
-    }
-
-    #[test]
-    fn test_multiple_heads() {
-        let mut list = TestRelativeList::new(10);
-
-        list.set_first(1, 2);
-        list.set_first(2, 3);
-
-        assert_eq!(list.get_first(1), 2);
-        assert_eq!(list.get_first(2), 3);
-    }
-}
-
-// =============================================================================
-// RelativeCircularLinkedList trait tests
-// =============================================================================
-
-#[cfg(test)]
-mod relative_circular_linked_list_tests {
-    use super::*;
-
-    const HEAD: usize = 1;
-
-    #[test]
-    fn test_attach_as_first_empty_list() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_first(HEAD, 2);
-
-        assert_eq!(list.get_first(HEAD), 2);
-        assert_eq!(list.get_last(HEAD), 2);
-        assert_eq!(list.get_size(HEAD), 1);
-        assert_eq!(list.get_previous(2), 2);
-        assert_eq!(list.get_next(2), 2);
-    }
-
-    #[test]
-    fn test_attach_as_first_non_empty_list() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_first(HEAD, 2);
-        list.attach_as_first(HEAD, 3);
-
-        assert_eq!(list.get_first(HEAD), 3);
-        assert_eq!(list.get_last(HEAD), 2);
-        assert_eq!(list.get_size(HEAD), 2);
-    }
-
-    #[test]
-    fn test_attach_as_last_empty_list() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_last(HEAD, 2);
-
-        assert_eq!(list.get_first(HEAD), 2);
-        assert_eq!(list.get_last(HEAD), 2);
-        assert_eq!(list.get_size(HEAD), 1);
-    }
-
-    #[test]
-    fn test_attach_as_last_non_empty_list() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_last(HEAD, 2);
-        list.attach_as_last(HEAD, 3);
-
-        assert_eq!(list.get_first(HEAD), 2);
-        assert_eq!(list.get_last(HEAD), 3);
-        assert_eq!(list.get_size(HEAD), 2);
-    }
-
-    #[test]
-    fn test_attach_before() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_first(HEAD, 2);
-        list.attach_as_last(HEAD, 3);
-        list.attach_before(HEAD, 3, 4);
-
-        assert_eq!(list.get_first(HEAD), 2);
-        assert_eq!(list.get_last(HEAD), 3);
-        assert_eq!(list.get_size(HEAD), 3);
-        assert_eq!(list.get_next(2), 4);
-        assert_eq!(list.get_next(4), 3);
-    }
-
-    #[test]
-    fn test_attach_before_first() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_first(HEAD, 2);
-        list.attach_before(HEAD, 2, 3);
-
-        assert_eq!(list.get_first(HEAD), 3);
-        assert_eq!(list.get_last(HEAD), 2);
-        assert_eq!(list.get_size(HEAD), 2);
-    }
-
-    #[test]
-    fn test_attach_after() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_first(HEAD, 2);
-        list.attach_as_last(HEAD, 3);
-        list.attach_after(HEAD, 2, 4);
-
-        assert_eq!(list.get_first(HEAD), 2);
-        assert_eq!(list.get_last(HEAD), 3);
-        assert_eq!(list.get_size(HEAD), 3);
-        assert_eq!(list.get_next(2), 4);
-        assert_eq!(list.get_next(4), 3);
-    }
-
-    #[test]
-    fn test_attach_after_last() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_first(HEAD, 2);
-        list.attach_after(HEAD, 2, 3);
-
-        assert_eq!(list.get_first(HEAD), 2);
-        assert_eq!(list.get_last(HEAD), 3);
-        assert_eq!(list.get_size(HEAD), 2);
-    }
-
-    #[test]
-    fn test_detach_single_element() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_first(HEAD, 2);
-        list.detach(HEAD, 2);
-
-        assert_eq!(list.get_first(HEAD), 0);
-        assert_eq!(list.get_last(HEAD), 0);
-        assert_eq!(list.get_size(HEAD), 0);
-        assert_eq!(list.get_previous(2), 0);
-        assert_eq!(list.get_next(2), 0);
-    }
-
-    #[test]
-    fn test_detach_first_element() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_first(HEAD, 2);
-        list.attach_as_last(HEAD, 3);
-        list.attach_as_last(HEAD, 4);
-
-        list.detach(HEAD, 2);
-
-        assert_eq!(list.get_first(HEAD), 3);
-        assert_eq!(list.get_last(HEAD), 4);
-        assert_eq!(list.get_size(HEAD), 2);
-    }
-
-    #[test]
-    fn test_detach_last_element() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_first(HEAD, 2);
-        list.attach_as_last(HEAD, 3);
-        list.attach_as_last(HEAD, 4);
-
-        list.detach(HEAD, 4);
-
-        assert_eq!(list.get_first(HEAD), 2);
-        assert_eq!(list.get_last(HEAD), 3);
-        assert_eq!(list.get_size(HEAD), 2);
-    }
-
-    #[test]
-    fn test_detach_middle_element() {
-        let mut list = TestRelativeList::new(10);
-
-        list.attach_as_first(HEAD, 2);
-        list.attach_as_last(HEAD, 3);
-        list.attach_as_last(HEAD, 4);
-
-        list.detach(HEAD, 3);
-
-        assert_eq!(list.get_first(HEAD), 2);
-        assert_eq!(list.get_last(HEAD), 4);
-        assert_eq!(list.get_size(HEAD), 2);
-        assert_eq!(list.get_next(2), 4);
-        assert_eq!(list.get_previous(4), 2);
-    }
-
-    #[test]
-    fn test_multiple_lists() {
-        const HEAD1: usize = 1;
-        const HEAD2: usize = 10;
-        let mut list = TestRelativeList::new(20);
-
-        list.attach_as_first(HEAD1, 2);
-        list.attach_as_last(HEAD1, 3);
-
-        list.attach_as_first(HEAD2, 11);
-        list.attach_as_last(HEAD2, 12);
-
-        assert_eq!(list.get_first(HEAD1), 2);
-        assert_eq!(list.get_last(HEAD1), 3);
-        assert_eq!(list.get_size(HEAD1), 2);
-
-        assert_eq!(list.get_first(HEAD2), 11);
-        assert_eq!(list.get_last(HEAD2), 12);
-        assert_eq!(list.get_size(HEAD2), 2);
-    }
-}
+use super::*;
 
 // =============================================================================
 // SizeBalancedTree trait tests
 // =============================================================================
 
-#[cfg(test)]
 mod size_balanced_tree_tests {
     use super::*;
 
@@ -1151,7 +379,6 @@ mod size_balanced_tree_tests {
 // IterativeSizeBalancedTree trait tests
 // =============================================================================
 
-#[cfg(test)]
 mod iterative_size_balanced_tree_tests {
     use super::*;
